requirements.txt
langgraph
langchain-ollama
langchain-core
pydantic
requests
feedparser
tavily-python
python-dotenv